Prevent Thread Oversubscription with threadpoolclt
In a cluster setting, thread oversubscription can lead to significant
performance degradation and resource contention for running the deconvolution
with scipy.optimize.  This commit addresses this issue by utilizing
the `threadpoolclt` library to limit the number of threads to 1.

This change ensures that each process uses only the allocated resources,
 preventing contention and improving overall cluster stability.